requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-requests==0.1.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cchardet==2.1.7
cffi==1.15.1
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pycares==4.3.0
pycparser==2.21
PyYAML==6.0
yarl==1.9.1
